scripts/python/requirements.txt
soccerdata>=1.7.0
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0
